+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## Unreleased
 
+- Allow the default configuration file's path to be configured at compilation time - [#28](https://github.com/emerick42/kairoi/pull/28)
 - Create a changelog file at the root of the project - [#27](https://github.com/emerick42/kairoi/pull/27)
 - Upgrade the minimum Rust version used to compile Kairoi to the current Rust version - [#26](https://github.com/emerick42/kairoi/pull/26)